Some more important changes
UUWAF now supports a session_dir argument which is used to set the path for sessions. Additionally, it is now probably the case that you only need specify the auth_dir and a new "base_dir" and all other paths will be automatically specified as per the (sorry Gordon) OPUS path. Obviously you can override any and all paths.



git-svn-id: svn+ssh://foss.ulster.ac.uk/svn/uuwaf/trunk@48 45a5810e-3be5-4c6a-b1ab-0eb843d56685